+++ Cod3rsGrowth.web/wwwroot/app/App.view.xml
@@ -3,12 +3,8 @@
       xmlns="sap.m"
       xmlns:mvc="sap.ui.core.mvc">
       <Shell>
-      <App
-            id="app"/>
-      <Button
-            id = "btnBemVindo"
-            text="{i18n>btnbemVindo}"
-            press=".onBemVindo"/>
+            <App
+                  id="app"/>
       </Shell>
