+++ Cod3rsGrowth.web/wwwroot/app/lista/Lista.view.xml
@@ -1,0 +1,13 @@
+<mvc:View
+	controllerName="ui5.anime.app.lista.Lista"
+	xmlns="sap.m"
+	xmlns:mvc="sap.ui.core.mvc">
+	<Page title="{i18n>homePageTitle}" class="sapUiResponsiveContentPadding">
+		<content>
+			<Button
+				id = "btnBemVindo"
+				text="{i18n>btnbemVindo}"
+				press=".onBemVindo"/>
+		</content>
+	</Page>
+</mvc:View>
