+++ Cod3rsGrowth.web/wwwroot/app/notFound/NotFound.view.xml
@@ -1,0 +1,11 @@
+<mvc:View
+controllerName="ui5.anime.app.notfound.NotFound"
+xmlns="sap.m"
+xmlns:mvc="sap.ui.core.mvc">
+<MessagePage
+    title="{i18n>NotFound}"
+    text="{i18n>NotFound.text}"
+    description="{i18n>NotFound.description}"
+    showNavButton="true"
+    navButtonPress="onNavBack"/>
+</mvc:View>
