+++ Cod3rsGrowth.web/wwwroot/i18n/i18n.properties
@@ -1,3 +1,9 @@
 btnbemVindo = Bem vindo
 bemVindoMsg = Bem vindo a lista de anime!
 btnFecharDialogo = OK
+
+NotFound=Página não encontrada
+NotFound.text=Desculpe, mas o recurso solicitado não está disponível.
+NotFound.description=Verifique o URL e tente novamente.
+
+homePageTitle = Home
