--- Cod3rsGrowth.web/wwwroot/index.js
@@ -1,9 +1,0 @@
-sap.ui.define([
-	"sap/ui/core/mvc/XMLView"
-], (XMLView) => {
-	"use strict";
-
-	XMLView.create({
-		viewName: "ui5.anime.app.App"
-	}).then((oView) => oView.placeAt("content"));
-});
